pyproject.toml
[tool.poetry]
name = "internet shop"
version = "0.1.0"
description = ""
authors = ["Andry2127 <andrybudnak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
flask-restful = "^0.3.10"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.37"
psycopg2 = "^2.9.10"
requests-html = "^0.10.0"
lxml = {extras = ["html-clean"], version = "^5.3.0"}
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
